+++ crates/vmi-os-windows/src/comps/peb.rs
@@ -1,21 +1,7 @@
 use vmi_core::{Architecture, Pa, Va, VmiDriver, VmiError, VmiState, VmiVa};
 
-use super::{macros::impl_offsets, process_parameters::WindowsProcessParameters};
+use super::{WindowsProcessParameters, WindowsWow64Kind, macros::impl_offsets};
 use crate::{ArchAdapter, WindowsOs};
-
-/// The address space type in a WoW64 process.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum WindowsWow64Kind {
-    /// Native address space.
-    Native = 0,
-
-    /// x86 (32-bit) address space under WoW64.
-    X86 = 1,
-    // Arm32 = 2,
-    // Amd64 = 3,
-    // ChpeX86 = 4,
-    // VsmEnclave = 5,
-}
 
 /// A Windows process environment block (PEB).
 ///