+++ crates/vmi-os-windows/src/lib.rs
@@ -81,7 +81,7 @@ pub use self::comps::{
     WindowsHandleTable, WindowsHandleTableEntry, WindowsImage, WindowsModule, WindowsObject,
     WindowsObjectAttributes, WindowsObjectHeaderNameInfo, WindowsObjectType, WindowsObjectTypeKind,
     WindowsPeb, WindowsProcess, WindowsProcessParameters, WindowsRegion, WindowsSectionObject,
-    WindowsSession, WindowsThread, WindowsTrapFrame, WindowsWow64Kind,
+    WindowsSession, WindowsTeb, WindowsThread, WindowsTrapFrame, WindowsWow64Kind,
 };
 
 /// VMI operations for the Windows operating system.