+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   foreign process (via `_KTHREAD.ApcStateIndex`)
 - `WindowsThread::saved_process()` - returns the thread's home process
   when attached to a foreign process (via `_KTHREAD.SavedApcState.Process`)
+- `WindowsThread::teb()` and `WindowsThread::native_teb()` - return the thread's TEB
 - `WindowsThread::trap_frame()` + `struct WindowsTrapFrame` - returns the thread's trap frame
 - `WindowsProcess::native_peb()` - returns the process's native PEB (via `_EPROCESS.Peb`)
 - `WindowsProcess::is_wow64()` - checks if the process is a WoW64 process (via `_EPROCESS.WoW64Process`)